Icesの方にペンギンの移動メソッド移動
